TPASPKT-685 Print subdirectories after upload
Reverted previous commit. Instead of generating unique random
directories, the client prints also subdirectories of directory of the
uploaded directory. Therefore, if user posts directories to root
directory, the identifiers of those directories are easily available,
and the risk of accidentally including the root directory to a dataset
is smaller.