TPASPKT-897 Refactor client error handling
Previously each client function had the same copy-pasted way to handle
errors by printing the JSON error to the screen and exit the system.
Handling the output and exiting the system is however the CLI's job, so
move those to the CLI code and remove the error handling boilerplate
from the client.

Changes also needlessly complicated file not found -error handling to
use this same generalised system. As this leaves all client functions
with a simple requests.raise_for_status(), use it as a session hook to
do the raise_for_status call automatically on every request and remove
the explicit raise_for_status calls from the client code as well.