TPASPKT-918 Do not retry requests
There has been a bug that the same POST request to upload-rest-api has
been sent twice, although the user has only executed one upload command
with the upload-client. As an attempt to fix this bug, disable retrying
failed requests.

As the bug could not be replicated, it is unclear whether this change
will help with the issue. If the bug persists, this commit should be
reverted.