TPASPKT-988 Remove faulty retry fix and log reqs
Remove faulty fix that prevents request retries. This has the
unfortunate effect of causing "remote disconnected" errors due to a race
condition between the client and the server, and it shouldn't also have
helped with the actual issue.

Add a random identifier to each non-idempotent request to help with
investigation of duplicate requests. Each identifier should only appear
once; if not, the request is duplicated by `requests` or one of the
underlying libraries.